+++ Runtime/CustomButtons/CanvasGroupVC.cs
@@ -11,6 +11,7 @@ namespace UI.CustomButtons
         [SerializeField] private CustomButtonState<float> _pressedAlpha;
         [SerializeField] private CustomButtonState<float> _selectedAlpha;
         [SerializeField] private CustomButtonState<float> _disabledAlpha;
+        [SerializeField] private CustomButtonState<float> _premiumAlpha;
 
         private float _startAlpha;
         private float _targetAlpha;
@@ -20,6 +21,7 @@ namespace UI.CustomButtons
         public override void OnPressed() => SetAlpha(_pressedAlpha.GetValue(_normalAlpha));
         public override void OnSelected() => SetAlpha(_selectedAlpha.GetValue(_normalAlpha));
         public override void OnDisabled() => SetAlpha(_disabledAlpha.GetValue(_normalAlpha));
+        public override void OnPremium() => SetAlpha(_premiumAlpha.GetValue(_normalAlpha));
 
         private void SetAlpha(float targetAlpha)
         {