+++ Runtime/ElementSelection/VisualModules/GameObjectToggleSelectionVisualModule.cs
@@ -8,9 +8,11 @@ namespace UI.ElementSelection.VisualModules
         [SerializeField] private GameObject _gameObject;
         [SerializeField] private bool _activateOnSelection = true;
         [SerializeField] private SelectionModuleState<bool> _hoveredValue;
+        [SerializeField] private SelectionModuleState<bool> _premiumValue;
 
         private bool _isSelected;
         private bool _isHovered;
+        private bool _isPremium;
         
         public void OnSelectionChanged(bool isSelected)
         {
@@ -24,15 +26,30 @@ namespace UI.ElementSelection.VisualModules
             UpdateActive();
         }
 
+        public void OnPremium()
+        {
+            _isPremium = true;
+            UpdateActive();
+        }
+
         private void UpdateActive()
         {
             if (_gameObject == null) return;
 
-            bool targetActive = _activateOnSelection ? _isSelected : !_isSelected;
-            
-            if (_isHovered && _hoveredValue.Enabled)
+            bool targetActive;
+
+            if (_isPremium && _premiumValue.Enabled)
             {
-                targetActive = _hoveredValue.Value;
+                targetActive = _premiumValue.Value;
+            }
+            else
+            {
+                targetActive = _activateOnSelection ? _isSelected : !_isSelected;
+                
+                if (_isHovered && _hoveredValue.Enabled)
+                {
+                    targetActive = _hoveredValue.Value;
+                }
             }
 
             _gameObject.SetActive(targetActive);